Add files via upload
In this program, ord('e') gets the ASCII value of 'e', and then it's converted to an integer. The bitwise AND operation is performed between the number nine and this integer value. The result is printed to the console.

When you run this program, you should see the result of the bitwise AND operation between the number nine and the mathematical constant 'e'.